refactor: commented out code for Dashboard view component to render without authentication
